--- src/slurm.rs
@@ -1,5 +1,0 @@
-use anyhow::{Result, bail};
-
-pub fn validate_slurm() -> Result<()> {
-    bail!("SLURM not configured")
-}
